pyproject.toml
[tool.poetry]
name = "cdk-gitbot"
version = "0.1.0"
description = "A gitbot powered by CDK"
authors = ["Damian Silbergleith <dsilberg@godaddy.com>"]

[tool.poetry.dependencies]
python = "^3.7"
"aws-cdk.core" = "1.90.1"
"aws-cdk.aws-apigateway" = "^1.90.1"
"aws-cdk.aws-dynamodb" = "1.90.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
